Don't close the fd passed to fdopendir
When creating a dir_list struct from an existing fd the system
takes ownership of that descriptor via the fdopendir syscall.
Don't close that descriptor after passing it to the
get_dir_entries_fd() funcion.

Add a function to create dir_list struct from a directory
descriptor similar to the system calls ending on at.