Make it possible to run chroot build from cli
Running the chroot build from cli required a chrooted shell to
run it from. Substitude the CMS_CONTENT and CMS_TEMPLATE variables
with the CMS_CHROOT prepended if a command line argument is
passed to the CGI. No longer needed to switch the build if
debugging from command line is required.